Basic grappling is implemented. Bushido can grab an AI(Left Trigger).If he misses the animation will play and he and his target move normally. If he is successful. Heavy Attack(Y) releases his target and Light Attack(X) will hit his target. He can hit his target 3 times before his target is automatically released. Judo Jaguar can counter grab the player if he attempts to grab him. He will release the player immediately afterward.
A new animation controller is being used called BushidoTestAnimator. It uses the same variables and animations as the old animator. It is wired up differently to better suit our needs and to make it easier to read.

To Do:
Block attacks
Throw grappled targets
Mechanic to allow Bushido to break away from grappling enemies(low priority)
Mechanic to allow enemies to break free of a grapple